Time: 13 ms (26.9%), Space: 19.7 MB (25.59%) - LeetHub

+++ 1938-minimum-operations-to-make-the-array-increasing/1938-minimum-operations-to-make-the-array-increasing.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int minOperations(vector<int>& nums) {
-        unsigned short numsSize = nums.size();
+        int numsSize = nums.size();
         int answer = 0;
 
-        for (unsigned short i = 0; i < numsSize - 1; ++i)
+        for (int i = 0; i < numsSize - 1; ++i)
         {
             if (nums[i] >= nums[i + 1])
             {
